src/app/questionnaire/questionnaire.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { QuestionnaireComponent } from './questionnaire.component';

describe('QuestionnaireComponent', () => {
  let component: QuestionnaireComponent;
  let fixture: ComponentFixture<QuestionnaireComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ QuestionnaireComponent ]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(QuestionnaireComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('Description should equall', () => {
    expect(component.description).toBe("Hey there. Interested in our product? Answer a few short questions and we'll get in touch with you!")
  });

  it('Description Note should equall', () => {
    expect(component.description_note).toBe("This will not take more than 5 minutes")
  });

});
